requirements.txt
aiogram==2.25.1
aiosqlite==0.18.0
phonenumbers==8.13.27
python-dotenv==1.0.0
pytz